mid conversion
