+++ lib/init.js
@@ -1,105 +1,37 @@
 'use strict';
 
-const _ = require('lodash'),
-  setup = require('./setup'),
+const setup = require('./setup'),
   bluebird = require('bluebird'),
-  es = require('./services/elastic'),
-  { indexWithPrefix } = require('./services/elastic-helpers'),
-  path = require('path'),
-  files = require('nymag-fs'),
-  acceptedIcons = ['icon.120x120.png', 'icon.180x180', 'icon.192x192.png'],
   lists = require('./lists'),
   routes = require('./routes'),
   db = require('./db'),
   bus = require('./bus');
-
-let log = require('./services/log').setup({file: __filename}),
-  SITES_INDEX;
-
-/**
- * Create the path to the media directory
- * for a site
- *
- * @param  {Object} site
- * @return {String}
- */
-function constructMediaPath(site) {
-  // For local development we want the user defined port
-  var port = (site.port === 80 || site.port === 443) ? '' : `:${site.port}`;
-
-  // Some sites, if they not in subdirectories of a main site, will not
-  // have an asset path. If no asset path then just use the site slug.
-  return `${site.host}${port}${site.path}/media/sites/${site.slug}/`;
-}
-
-
-/**
- * Populate the internal sites index with all the
- * sites included in the instance.
- *
- * @return {Promise}
- */
-function generateSitesIndexOps() {
-  var ops = [], // To be populated with Elastic ops for each site
-    sites = setup.options.sites.sites(); // Grab the sites using the passed in options from setup
-
-  SITES_INDEX = indexWithPrefix('sites'); // note: this is exported for page list
-
-  _.each(sites, function (site) {
-    var mediaDir = path.resolve(process.cwd(), site.assetDir, 'media', 'sites', site.slug),
-      mediaDirFiles = files.getFiles(mediaDir),
-      icons = _.intersection(acceptedIcons, mediaDirFiles); // https://lodash.com/docs/4.17.2#intersection
-
-    ops.push({
-      index: {
-        _index: SITES_INDEX,
-        _type: '_doc',
-        _id: site.slug
-      }
-    }, {
-      name: site.name,
-      slug: site.slug,
-      host: site.host,
-      path: site.path,
-      port: site.port,
-      protocol: site.protocol,
-      assetDir: site.assetDir,
-      assetPath: site.assetPath,
-      mediaPath: `${constructMediaPath(site)}`,
-      siteIcon: `${icons[0]}`
-    });
-  });
-
-  return ops;
-}
+let log = require('./services/log').setup({file: __filename});
 
 /**
  * Add all the sites to the Sites index.
  *
- * @returns {Promise}
+ * @param {Object} options
+ * @returns {Function}
  */
-function onInit(router, storage) {
-  return es.batch(module.exports.generateSitesIndexOps())
-    .then(() => db(storage))
-    .then(() => routes(router))
-    .then(bus)
-    .then(lists)
-    .catch(err => {
-      log('fatal', `${err.message}`, {
-        stack: err.stack
+function onInit(options) {
+  return (router, storage, busPublish, sites) => {
+    return setup(options)
+      .then(() => db(storage))
+      .then(() => routes(router))
+      .then(bus)
+      .then(() => lists(sites))
+      .catch(err => {
+        log('fatal', `${err.message}`, {
+          stack: err.stack
+        });
+        return bluebird.reject(err);
       });
-      return bluebird.reject(err);
-    });
-}
-
-// For testing
-function setSites(sitesObj) {
-  setup.options.sites = sitesObj;
+  };
 }
 
 module.exports = onInit;
+
 // For testing
-module.exports.constructMediaPath = constructMediaPath;
-module.exports.generateSitesIndexOps = generateSitesIndexOps;
-module.exports.setSites = setSites;
-module.exports.setLog = fakeLogger => { log = fakeLogger; };
+module.exports.setSites = mock => setup.options.sites = mock;
+module.exports.setLog = mock => log = mock;
