+++ lib/lists/index.js
@@ -5,7 +5,8 @@ const h = require('highland'),
   { subscribe } = require('../streams'),
   pageList = require('./page-list'),
   layoutList = require('./layout-list'),
-  userList = require('./user-list');
+  userList = require('./user-list'),
+  sitesList = require('./sites-list');
 var log = require('../services/log').setup({ file: __filename });
 
 // For meta updates
@@ -54,10 +55,12 @@ function handleMetaSave({ uri, data }) {
 /**
  *
  */
-function setup() {
+function setup(sitesService) {
   pageList.setPagesIndex();
   layoutList.setLayoutsIndex();
   userList.setUserIndex();
+
+  return sitesList(sitesService);
 }
 
 module.exports = setup;
