+++ lib/setup.js
@@ -7,7 +7,8 @@ const _ = require('lodash'),
 
 var mappings = {},
   settings = {},
-  handlers = {};
+  handlers = {},
+  ALREADY_RUN = false;
 
 /**
  * Load the mappings from the mappings folder.
@@ -18,7 +19,7 @@ var mappings = {},
 function loadMappingConfiguration(mappingDir) {
   const list = files.getFiles(mappingDir);
 
-  _.each(list, function (filename) {
+  _.each(list, filename => {
     const mappingName = filename.split('.')[0],
       mapping = files.getYaml(path.join(mappingDir, mappingName));
 
@@ -55,15 +56,6 @@ function loadHandlers(handlersDir) {
   module.exports.handlers = handlers;
 }
 
-/**
- * [setPrefix description]
- * @param {String} prefix
- */
-function setPrefix(prefix) {
-  if (prefix && _.isString(prefix)) {
-    module.exports.prefix = prefix;
-  }
-}
 
 /**
  * Setup all necessary parts else necessary to use Elastic
@@ -73,25 +65,30 @@ function setPrefix(prefix) {
  * @return {Promise}
  */
 function setup(options) {
+  if (ALREADY_RUN) { // Run this once
+    return Promise.resolve();
+  }
   // Make the options accessible
   module.exports.options = options;
 
   // Grab the mappings from the specified directory
-  if (options.mappings) {
+  if (!options.skipMappings) {
     loadMappingConfiguration(path.resolve(process.cwd(), './search/mappings'));
     loadMappingConfiguration(path.resolve(__dirname, '../mappings'));
   }
 
-  // Load handlers
-  if (options.handlers) {
+  if (!options.skipHandlers) {
     loadHandlers(path.resolve(process.cwd(), './search/handlers'));
   }
 
   // Set prefix
-  setPrefix(options.prefix);
+  module.exports.prefix = process.env.ELASTIC_PREFIX || '';
 
   // Setup the ES client
   es.setup(module.exports.host);
+
+  // Ok, we need to make sure we don't run this again
+  ALREADY_RUN = true;
 
   // Validate and create the indices
   return es.validateIndices(mappings, settings, module.exports.prefix);
